+++ demo/src/main/java/com/example/demo/servis/EnterpriseServiceInterface.java
@@ -7,10 +7,10 @@ import org.springframework.data.domain.Page;
 import com.example.demo.model.Enterprise;
 
 public interface EnterpriseServiceInterface {
-	Enterprise findOne(Integer enterprise_id);
+	Enterprise findOne(Long enterprise_id);
 	List <Enterprise> findAll();
 	Enterprise save(Enterprise enterprise);
-	void remove(Integer enterprise_id);
+	void remove(Long enterprise_id);
 	Enterprise findByName(String name);
 	
 	Page<Enterprise> findAll(int pageNo, int pageSize);
