+++ demo/src/main/java/com/example/demo/servis/PricelistServiceInterface.java
@@ -9,13 +9,13 @@ public interface PricelistServiceInterface {
 	
 	Pricelist findByDateFrom (Date dateFrom);
 
-	Pricelist findOne(Integer pricelist_id);
+	Pricelist findOne(Long pricelist_id);
 	
 	List<Pricelist> findAll();
 	
 	Pricelist save(Pricelist pricelist);
 	
-	void remove(Integer pricelist_id);
+	void remove(Long pricelist_id);
 	
 //	Pricelist findByTotalPrice(Integer total_price);
